4089: Reverted getSlugFromPath and moved getCategoryTiles to categories.ts
