+++ shared/api/models/DocumentModel.ts
@@ -1,9 +1,8 @@
 import { decodeHTML } from 'entities'
 import { DateTime } from 'luxon'
 
+import { getSlugFromPath } from '../../utils'
 import normalizePath from '../../utils/normalizePath'
-
-export const getSlugFromPath = (path: string): string => path.split('/').pop() ?? ''
 
 class DocumentModel {
   _path: string
