+++ shared/api/models/ExtendedDocumentModel.ts
@@ -1,7 +1,8 @@
 import { DateTime } from 'luxon'
 
+import { getSlugFromPath } from '../../utils'
 import isEqual from '../../utils/isEqual'
-import DocumentModel, { getSlugFromPath } from './DocumentModel'
+import DocumentModel from './DocumentModel'
 
 class ExtendedDocumentModel extends DocumentModel {
   _thumbnail: string | null
