+++ shared/utils/__tests__/index.spec.ts
@@ -2,14 +2,14 @@ import { DateTime } from 'luxon'
 
 import { CategoriesMapModelBuilder } from '../../api'
 import CategoryModel from '../../api/models/CategoryModel'
-import { getSlugFromPath } from '../../api/models/DocumentModel'
 import OfferModel from '../../api/models/OfferModel'
 import TileModel from '../../models/TileModel'
+import { getCategoryTiles } from '../categories'
 import {
   addSubdomain,
   formatDateICal,
   getGenericLanguageCode,
-  getCategoryTiles,
+  getSlugFromPath,
   safeParseInt,
   segmentText,
 } from '../index'
